Split NRS IFU spec3 regtest in two tests
